atualização do projeto - controllers e endpoints prontos

+++ jobs/src/main/java/br/ufsm/csi/jobs/service/EmpresaService.java
@@ -1,5 +1,6 @@
 package br.ufsm.csi.jobs.service;
 
+import br.ufsm.csi.jobs.error.EmpresaNotFoundException;
 import br.ufsm.csi.jobs.model.Empresa;
 import br.ufsm.csi.jobs.repo.EmpresaRepo;
 import br.ufsm.csi.jobs.repo.VagaRepo;
@@ -12,7 +13,6 @@ import java.util.Optional;
 
 @Service
 public class EmpresaService {
-
     private final EmpresaRepo empresaRepo;
     private final VagaRepo vagaRepo;
 
@@ -34,11 +34,16 @@ public class EmpresaService {
         return empresaRepo.findAll();
     }
 
-    public void deleteEmpresaById(Long id) {
-        empresaRepo.deleteById(id);
+    public void deleteEmpresaById(Long id) throws EmpresaNotFoundException {
+        if (empresaRepo.existsById(id)) {
+            empresaRepo.deleteById(id);
+        } else {
+            throw new EmpresaNotFoundException();
+        }
     }
 
     @Transactional
     public void deleteVagasByEmpresaId(Long id) {
-        vagaRepo.deleteByEmpresaId(id);    }
+        vagaRepo.deleteByEmpresaId(id);
+    }
 }
